% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saleOrderTable.R
\name{saleOrderTable}
\alias{saleOrderTable}
\title{查看即时库存更新报表}
\usage{
saleOrderTable(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
查看即时库存更新报表
}
\examples{
saleOrderTable()
}
